feat(admin): add deletion restriction for master user - move MASTER_ID constant into variable.py - rename user into user_id in DELETE /user endpoint
