+++ api/endpoints/admin/users.py
@@ -44,19 +44,19 @@ async def create_user(
 
 
 @router.delete(
-    path=EndpointRoute.ADMIN_USERS + "/{user}",
+    path=EndpointRoute.ADMIN_USERS + "/{user_id}",
     dependencies=[Security(dependency=AccessController(permissions=[PermissionType.ADMIN]))],
     status_code=204,
 )
 async def delete_user(
     request: Request,
-    user: int = Path(description="The ID of the user to delete."),
+    user_id: int = Path(description="The ID of the user to delete."),
     postgres_session: AsyncSession = Depends(get_postgres_session),
 ) -> Response:
     """
     Delete a user.
     """
-    await global_context.identity_access_manager.delete_user(postgres_session=postgres_session, user_id=user)
+    await global_context.identity_access_manager.delete_user(postgres_session=postgres_session, user_id=user_id)
 
     return Response(status_code=204)
