+++ api/tests/integration/postgres/test_postgresrouterrepository.py
@@ -13,6 +13,7 @@ from api.sql.models import Provider as ProviderTable
 from api.sql.models import Router as RouterTable
 from api.sql.models import RouterAlias as RouterAliasTable
 from api.tests.integration.factories import OrganizationSQLFactory, RouterSQLFactory, UserSQLFactory
+from api.utils.variables import MASTER_ID
 
 
 def to_router_domain(router_sql, aliases: list[str] | None = None) -> Router:
